Allow specifying a starting-offset in message lists.
We now search for all valid UIDs when retrieving the message list,
even with 50,000 messages in a folder this is fast.  It might be
possible to avoid this but I ran into issues where the sequence
numbers had gaps - they are NOT contiguous.

Anyway when we have retrieved all the sequence numbers that are
valid we can choose 50 to retrieve, from the given offset.

This updates #9 to make it possible to page in the future:

TODO

* Return min/max from the function.
* Allow the template to show a next/prev button.